// web/src/api/models/TypesFriendInterface.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Routier API
 * Configuration and monitoring API for a Routier node.
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface TypesFriendInterface
 */
export interface TypesFriendInterface {
    /**
     * 
     * @type {Array<string>}
     * @memberof TypesFriendInterface
     */
    addresses?: Array<string>;
    /**
     * 
     * @type {string}
     * @memberof TypesFriendInterface
     */
    name: string;
}

/**
 * Check if a given object implements the TypesFriendInterface interface.
 */
export function instanceOfTypesFriendInterface(value: object): value is TypesFriendInterface {
    if (!('name' in value) || value['name'] === undefined) return false;
    return true;
}

export function TypesFriendInterfaceFromJSON(json: any): TypesFriendInterface {
    return TypesFriendInterfaceFromJSONTyped(json, false);
}

export function TypesFriendInterfaceFromJSONTyped(json: any, ignoreDiscriminator: boolean): TypesFriendInterface {
    if (json == null) {
        return json;
    }
    return {
        
        'addresses': json['addresses'] == null ? undefined : json['addresses'],
        'name': json['name'],
    };
}

export function TypesFriendInterfaceToJSON(json: any): TypesFriendInterface {
    return TypesFriendInterfaceToJSONTyped(json, false);
}

export function TypesFriendInterfaceToJSONTyped(value?: TypesFriendInterface | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'addresses': value['addresses'],
        'name': value['name'],
    };
}